Move model_for_depl fixture
